add more and less tasks
The number simulation now contains 5 types of tasks:

- predecessor to
- successor to
- equal to
- greater than
- less than

Also, I fixed a bug in the library. Previously, the basic-level object
(in this case, the integer 1) was listed in the wrong library and wasn't
available for number learning. Now, it's been moved to the number
learning library and is available.

Also, I decreased the learning threshold in the test script in an
attempt to increase the number of productions that are available for the
grammar to recombine.

Finally, I changed the "bag" functions in the library to make use of the
built-in set functions defined over lists.