(Det, Noun, world) -> world, 1 task, no learning
I've changed the task architecture to be from an ad-hoc triple to a set
(the world) to be from a Determiner, Noun, and World triple to a world.
I've still been using a single task, though, which I now think is a
mistake. Primarily, I think the drive to compress should come from E.C.,
not the task structure. We still encounter tasks as ad-hoc, and we don't
necessarily know how to respond coherently to them. Compression should
teach us to do that and generalize appropriately. So, I'm recording this
for posterity and moving to re-introduce multiple tasks. I think that,
as long as we forego the ability to ignore arguments (no K combinator),
this approach has a decent shot at working.