further experiments with Number, new files written
There's a number of changes represented here, but perhaps the most
useful are the new files I've started writing out to look at how the
grammar and performance of the system evolves over time.

I'm planning to move forward with a new type for the task, so this is
the last commit (for now) using triples to represent the query. I'm
trying multiple arguments now.